Remove redundant type casts

+++ src/main/java/rv/comm/rcssserver/LogAnalyzerThread.java
@@ -114,7 +114,7 @@ public class LogAnalyzerThread extends Thread
 
 		if (scoringTeam != -1) {
 			int frame = logfile.getCurrentFrame();
-			int goalWindowFrames = (int) Math.round((1 / stepSize) * LogPlayer.GOAL_WINDOW_SECONDS);
+			int goalWindowFrames = Math.round((1 / stepSize) * LogPlayer.GOAL_WINDOW_SECONDS);
 			int viewFrame = Math.max(0, frame - goalWindowFrames);
 			callback.goalFound(new Goal(frame, viewFrame, scoringTeam));
 		}
@@ -143,7 +143,7 @@ public class LogAnalyzerThread extends Thread
 				float stepSize = time - lastTime;
 				// estimate total number of frames
 				float halfTime = world.getGameState().getHalfTime();
-				int numFrames = (int) Math.round(((1 / stepSize) * halfTime) + numPauseFrames);
+				int numFrames = Math.round(((1 / stepSize) * halfTime) + numPauseFrames);
 				callback.stepSizeFound(stepSize, numFrames);
 				this.stepSize = stepSize;
 			}
